基本宛城

+++ app/src/main/java/com/haife/app/nobles/spirits/kotlin/mvp/contract/MainContract.java
@@ -35,9 +35,9 @@ public interface MainContract {
 
     //Model层定义接口,外部只需关心Model返回的数据,无需关心内部细节,即是否使用缓存
     interface Model extends IModel {
-        Observable<BaseResponse<LoginInfoBean>> loginInfo(int uId,  String sId);
+        Observable<BaseResponse<LoginInfoBean>> loginInfo(long uId,  String sId);
 
-        Observable<BaseResponse> addFriend(int uId, String sId, int friendUid,  String remark);
-        Observable<BaseResponse<MyGroup>> addGroup(int uId, String sId,int groupId);
+        Observable<BaseResponse> addFriend(long uId, String sId, long friendUid,  String remark);
+        Observable<BaseResponse<MyGroup>> addGroup(long uId, String sId,long groupId);
     }
 }
