+++ app/src/main/java/com/haife/app/nobles/spirits/kotlin/mvp/model/bean/R_AddGroupBean.java
@@ -1,11 +1,11 @@
 package com.haife.app.nobles.spirits.kotlin.mvp.model.bean;
 
 public class R_AddGroupBean {
-    int UID;
+    long UID;
     String SID;
-    int friendUid;
+    long friendUid;
 
-    public R_AddGroupBean(int UID, String SID, int friendUid) {
+    public R_AddGroupBean(long UID, String SID, long friendUid) {
         this.UID = UID;
         this.SID = SID;
         this.friendUid = friendUid;
