+++ app/src/main/java/com/haife/app/nobles/spirits/kotlin/mvp/model/bean/R_FriendMsgBean.java
@@ -3,9 +3,9 @@ package com.haife.app.nobles.spirits.kotlin.mvp.model.bean;
 public class R_FriendMsgBean {
     CharSequence msgContent;
    int msgType;
-   int receiverUid;
+    long receiverUid;
 
-    public R_FriendMsgBean(CharSequence msgContent, int msgType, int receiverUid) {
+    public R_FriendMsgBean(CharSequence msgContent, int msgType, long receiverUid) {
         this.msgContent = msgContent;
         this.msgType = msgType;
         this.receiverUid = receiverUid;
