+++ app/src/main/java/com/haife/app/nobles/spirits/kotlin/mvp/model/bean/R_loginBean.java
@@ -1,10 +1,10 @@
 package com.haife.app.nobles.spirits.kotlin.mvp.model.bean;
 
 public class R_loginBean {
-    int uid;
+    long uid;
     String pwd;
 
-    public R_loginBean(int uid, String pwd) {
+    public R_loginBean(long uid, String pwd) {
         this.uid = uid;
         this.pwd = pwd;
     }
